tee/python/requirements.txt
pycryptodome>=3.20
coincurve>=20.0
eth-abi>=5.0
